add getCurrentPokemon func
