+++ src/redux/sagas/rootSaga.js
@@ -1,7 +1,6 @@
 import { all } from 'redux-saga/effects';
 
-import { watchFetchPokemons } from './fetchDataSaga';
-import { watchGetCurrentPokemon } from './getCurrentPokemonSaga';
+import { watchFetchPokemons, watchGetCurrentPokemon } from './fetchDataSaga';
 
 export default function* rootSaga() {
   yield all([
